+++ src/Resolver/CallResolver/CalledClassResolver.php
@@ -10,6 +10,8 @@ use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\StaticCall;
 use PhpParser\Node\Expr\Variable;
 use PHPStan\Analyser\Scope;
+use PHPStan\BetterReflection\BetterReflection;
+use PHPStan\BetterReflection\Reflector\Exception\IdentifierNotFound;
 use PHPStan\Type\ObjectType;
 
 final class CalledClassResolver
@@ -55,5 +57,32 @@ final class CalledClassResolver
             $callerType = $scope->getType($node->var);
             return $callerType instanceof ObjectType ? $callerType->getClassName() : null;
         }
+    }
+
+    public function resolveDeclaring(Node $node, Scope $scope): ?string
+    {
+        if (!$node instanceof MethodCall && !$node instanceof StaticCall) {
+            return null;
+        }
+
+        $calledClassName = $this->resolve($node, $scope);
+        $calledMethodName = $this->nameResolver->resolve($node);
+
+        if ($calledClassName === null || $calledMethodName === null || $calledMethodName === '') {
+            return null;
+        }
+
+        try {
+            $reflectionClass = (new BetterReflection())->reflector()->reflectClass($calledClassName);
+        } catch (IdentifierNotFound $e) {
+            return null;
+        }
+
+        $reflectionMethod = $reflectionClass->getMethod($calledMethodName);
+        if ($reflectionMethod === null) {
+            return null;
+        }
+
+        return $reflectionMethod->getDeclaringClass()->getName();
     }
 }